more fixes and liger-type chunked loss
